Relax Encoder Bounds
This PR modifies the `Encode + Decode` bounds on parameter types for the
various encoders, allowing one to implement _just_ the trait that is
needed on special types.

For example, if you implement `Encode` for a custom type for
ABI-encoding funtion parameters, and never try to decode the function
parameter, then you can still use it with the `FunctionEncoder` type.